refactor(animation): Extract AnimationLoop from GameCanvas
Moves the core animation and rendering logic out of the GameCanvas React component and into a dedicated AnimationLoop class. This improves separation of concerns and prepares the animation logic for direct unit testing, as outlined in Step 3.1 of the testing plan.
